+++ src/spflow/torch/sampling/general/nodes/leaves/parametric/gamma.py
@@ -59,6 +59,6 @@ def sample(
 
     sampling_ids = marg_ids & instance_ids_mask.bool().to(leaf.alpha_aux.device)
 
-    data[sampling_ids, leaf.scope.query] = leaf.dist.sample((sampling_ids.sum(),)).to(leaf.alpha_aux.device)
+    data[sampling_ids, leaf.scope.query] = leaf.dist.sample((sampling_ids.sum(),)).double().to(leaf.alpha_aux.device)
 
     return data
